requirements.txt
streamlit
pillow
numpy
pandas
tensorflow==2.0.0
setuptools==41.0.0
keras
